+++ training/core/predict.py
@@ -30,12 +30,12 @@ import pandas as pd
 import polars as pl
 from pyutilz.strings import slugify
 
-from .configs import TargetTypes
-from .extractors import FeaturesAndTargetsExtractor
-from .io import load_mlframe_model
-from .pipeline import prepare_df_for_catboost
-from .utils import drop_columns_from_dataframe
-from .core_utils import (
+from ..configs import TargetTypes
+from ..extractors import FeaturesAndTargetsExtractor
+from ..io import load_mlframe_model
+from ..pipeline import prepare_df_for_catboost
+from ..utils import drop_columns_from_dataframe
+from .utils import (
     DEFAULT_PROBABILITY_THRESHOLD,
     _drop_cols_df,
     _setup_model_directories,